New database class
init_db(), and populate_db() are __init__(self) and populate(self) respectively. Also added function for adding a new row to the assignments table using the assignments class. I expected we would have a barebones website up by now, but that is not the case. Thinking about Django and Flash Web Frameworks. I have used Django in a past class and Ill see about making my next commit a converted version of that for a possible website use. We can decide if thats the direction we want to go and merge that commit, or look into something else.