Update notification.py
Updated documentation and added NotificationDelegate class decoration. The delegate will go out when Notification.send_notification() is called. The Notification.send_notification() is currently called from a separate thread that is always checking if the notification date is the current date. The AssignmentManager should attach to this delegate and call Notification.stop_tick() if the assignment is non-recurring, otherwise the date should be set to 1 week (should this be a variable? How often the assignment recurrs?) from the current date.